rename: 이미지 url 생성 util 클래스 패키지 위치 및 네이밍 변경
