+++ backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/response/CreateAuctionResponse.java
@@ -1,8 +1,8 @@
 package com.ddang.ddang.auction.presentation.dto.response;
 
 import com.ddang.ddang.auction.application.dto.CreateInfoAuctionDto;
-import com.ddang.ddang.image.util.ImageBaseUrl;
-import com.ddang.ddang.image.util.ImageUrlBuilder;
+import com.ddang.ddang.image.presentation.util.ImageBaseUrl;
+import com.ddang.ddang.image.presentation.util.ImageUrlCalculator;
 
 public record CreateAuctionResponse(
         Long id,
@@ -26,6 +26,6 @@ public record CreateAuctionResponse(
     }
 
     private static String convertAuctionImageUrl(final Long id) {
-        return ImageUrlBuilder.calculate(ImageBaseUrl.AUCTION, id);
+        return ImageUrlCalculator.calculate(ImageBaseUrl.AUCTION, id);
     }
 }
