+++ backend/ddang/src/main/java/com/ddang/ddang/user/presentation/dto/ReadUserResponse.java
@@ -1,7 +1,7 @@
 package com.ddang.ddang.user.presentation.dto;
 
-import com.ddang.ddang.image.util.ImageBaseUrl;
-import com.ddang.ddang.image.util.ImageUrlBuilder;
+import com.ddang.ddang.image.presentation.util.ImageBaseUrl;
+import com.ddang.ddang.image.presentation.util.ImageUrlCalculator;
 import com.ddang.ddang.user.application.dto.ReadUserDto;
 
 public record ReadUserResponse(String name, String profileImage, double reliability) {
@@ -15,6 +15,6 @@ public record ReadUserResponse(String name, String profileImage, double reliabil
     }
 
     private static String convertImageUrl(final Long id) {
-        return ImageUrlBuilder.calculate(ImageBaseUrl.USER, id);
+        return ImageUrlCalculator.calculate(ImageBaseUrl.USER, id);
     }
 }
